+++ bind/backend.go
@@ -38,12 +38,12 @@ type ContractCaller interface {
 // to the transactor to decide.
 type ContractTransactor interface {
 
-	// EstimateGasAndCollateral tries to estimate the gas and storage collateral
+	// EstimateGas tries to estimate the gas and storage collateral
 	// needed to execute a specific transaction based on the current pending state
 	// of the backend blockchain. There is no guarantee that this is the true gas
 	// limit requirement as other transactions may be added or removed by miners,
 	// but it should provide a basis for setting a reasonable default.
-	EstimateGasAndCollateral(request types.CallRequest, epoch ...*types.Epoch) (estimat types.Estimate, err error)
+	EstimateGas(request types.CallRequest, epoch ...*types.Epoch) (estimat types.Estimate, err error)
 	// SendTransaction injects the transaction into the pending pool for execution.
 	SendTransaction(tx types.UnsignedTransaction) (types.Hash, error)